+++ skrub/__init__.py
@@ -6,8 +6,8 @@ from pathlib import Path as _Path
 
 from . import selectors
 from ._agg_joiner import AggJoiner, AggTarget
-from ._apply_to_cols import ApplyOnEachCol
-from ._apply_to_frame import ApplySubFrame
+from ._apply_on_each_col import ApplyOnEachCol
+from ._apply_sub_frame import ApplySubFrame
 from ._column_associations import column_associations
 from ._config import config_context, get_config, set_config
 from ._data_ops import (
