+++ skrub/_wrap_transformer.py
@@ -1,5 +1,5 @@
-from ._apply_to_cols import ApplyOnEachCol
-from ._apply_to_frame import ApplySubFrame
+from ._apply_on_each_col import ApplyOnEachCol
+from ._apply_sub_frame import ApplySubFrame
 from ._single_column_transformer import is_single_column_transformer
 from .selectors import make_selector
